refactor: Extract styles
